test(e2e): split the sweeper into its own project, out of the normal suite
delete.spec.js mixed a real functional test ("Empty out open editors") with
the scheduled sweeper ("Delete multiple old pages"). The sweeper was kept out
of normal runs only via --grep-invert, which - like the old teardown - breaks
under bare npx / IDE runs.

Move the sweeper to sweeper.spec.js, testIgnore it from the browser projects,
and run it via a dedicated 'sweeper' project (test:cleanup --project=sweeper).
delete.spec.js keeps just the functional test. Drop the now-unneeded
--grep-invert from the run scripts.
