+++ test/e2e/package.json
@@ -2,11 +2,11 @@
   "name": "da-live_e2e_test",
   "version": "1.0.0",
   "scripts": {
-    "test": "playwright test --project=chromium --grep-invert \"Delete multiple old pages\"",
+    "test": "playwright test --project=chromium",
     "test:install": "playwright install",
-    "test:all": "playwright test --project=chromium --project=firefox --project=webkit --grep-invert \"Delete multiple old pages\"",
-    "test:nonauth": "SKIP_AUTH=true playwright test --project=chromium tests/*.spec.js --grep-invert \"Delete multiple old pages\"",
-    "test:cleanup": "playwright test --project=chromium tests/delete.spec.js --grep \"Delete multiple old pages\" --retries=1",
+    "test:all": "playwright test --project=chromium --project=firefox --project=webkit",
+    "test:nonauth": "SKIP_AUTH=true playwright test --project=chromium tests/*.spec.js",
+    "test:cleanup": "playwright test --project=sweeper --retries=1",
     "test:ui": "playwright test --project=chromium --ui",
     "test:debug": "playwright test --project=chromium --debug",
     "test:report": "playwright show-report"
